feat: separate types
