+++ .eslintignore
@@ -4,3 +4,4 @@ content
 node_modules
 storybook-static
 client-app/core/api/graphql/**/types.ts
+client-app/modules/**/api/types.ts
