+++ client-app/modules/quotes/api/mutations/approveQuoteRequest/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import mutationDocument from "./approveQuoteRequestMutation.graphql";
-import type { Mutations, MutationsApproveQuoteRequestArgs, ApproveQuoteResultType } from "@/core/api/graphql/types";
+import type { Mutations, MutationsApproveQuoteRequestArgs, ApproveQuoteResultType } from "../../types";
 
 export async function approveQuoteRequest(payload: MutationsApproveQuoteRequestArgs): Promise<ApproveQuoteResultType> {
   const { data } = await graphqlClient.mutate<
