+++ client-app/modules/quotes/api/mutations/changeQuoteComment/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import mutationDocument from "./changeQuoteCommentMutation.graphql";
-import type { Mutations, MutationsChangeQuoteCommentArgs, QuoteType } from "@/core/api/graphql/types";
+import type { Mutations, MutationsChangeQuoteCommentArgs, QuoteType } from "../../types";
 
 export async function changeQuoteComment(payload: MutationsChangeQuoteCommentArgs): Promise<QuoteType> {
   const { data } = await graphqlClient.mutate<
