+++ client-app/modules/quotes/api/mutations/changeQuoteItemQuantity/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import mutationDocument from "./changeQuoteItemQuantityMutation.graphql";
-import type { Mutations, MutationsChangeQuoteItemQuantityArgs, QuoteType } from "@/core/api/graphql/types";
+import type { Mutations, MutationsChangeQuoteItemQuantityArgs, QuoteType } from "../../types";
 
 export async function changeQuoteItemQuantity(payload: MutationsChangeQuoteItemQuantityArgs): Promise<QuoteType> {
   const { data } = await graphqlClient.mutate<
