+++ client-app/modules/quotes/api/mutations/createQuote/index.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "@/core/api/graphql/composables";
-import { CreateQuoteDocument } from "@/core/api/graphql/types";
 import { globals } from "@/core/globals";
+import { CreateQuoteDocument } from "../../types";
 
 export function useCreateQuoteMutation() {
   const { storeId, userId, currencyCode, cultureName } = globals;
