+++ client-app/modules/quotes/api/mutations/removeQuoteItem/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import mutationDocument from "./removeQuoteItemMutation.graphql";
-import type { Mutations, MutationsRemoveQuoteItemArgs, QuoteType } from "@/core/api/graphql/types";
+import type { Mutations, MutationsRemoveQuoteItemArgs, QuoteType } from "../../types";
 
 export async function removeQuoteItem(payload: MutationsRemoveQuoteItemArgs): Promise<QuoteType> {
   const { data } = await graphqlClient.mutate<
