+++ client-app/modules/quotes/api/mutations/submitQuoteRequest/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import mutationDocument from "./submitQuoteRequestMutation.graphql";
-import type { Mutations, MutationsSubmitQuoteRequestArgs, QuoteType } from "@/core/api/graphql/types";
+import type { Mutations, MutationsSubmitQuoteRequestArgs, QuoteType } from "../../types";
 
 export async function submitQuoteRequest(payload: MutationsSubmitQuoteRequestArgs): Promise<QuoteType> {
   const { data } = await graphqlClient.mutate<
