+++ client-app/modules/quotes/api/mutations/updateQuoteAttachments/index.ts
@@ -1,6 +1,6 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import updateQuoteAttachmentsDocument from "./updateQuoteAttachments.graphql";
-import type { UpdateQuoteAttachmentsMutation, UpdateQuoteAttachmentsMutationVariables } from "@/core/api/graphql/types";
+import type { UpdateQuoteAttachmentsMutation, UpdateQuoteAttachmentsMutationVariables } from "../../types";
 
 export async function updateQuoteAttachments(
   quoteId: string,
