+++ client-app/modules/quotes/api/queries/getQuotes/index.ts
@@ -1,7 +1,7 @@
 import { graphqlClient } from "@/core/api/graphql/client";
 import { globals } from "@/core/globals";
 import getQuotesQueryDocument from "./getQuotesQuery.graphql";
-import type { Query, QueryQuotesArgs, QuoteConnection } from "@/core/api/graphql/types";
+import type { Query, QueryQuotesArgs, QuoteConnection } from "../../types";
 
 export async function getQuotes(payload?: QueryQuotesArgs): Promise<QuoteConnection> {
   const { storeId, userId, cultureName, currencyCode } = globals;
